# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Shop Catalog API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublic"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/block:
    post:
      operationId: blockUser
      tags: [users]
      summary: Block or unblock user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
  /users/{id}/force-logout:
    post:
      operationId: forceLogoutUser
      tags: [users]
      summary: Force logout a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
...
components:
  schemas:
...
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, manager, user]
        isBlocked:
          type: boolean
        hasActiveSession:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, username, role, isBlocked, hasActiveSession, createdAt]
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, manager, user]
      required: [username, password, role]
    UpdateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, manager, user]
    BlockUserBody:
      type: object
      properties:
        isBlocked:
          type: boolean
      required: [isBlocked]
...